Validation complete in 50%

+++ WalletAPP/Categories.xaml.cs
@@ -38,39 +38,52 @@ namespace WalletAPP
 
         private void removeIncomeCategory_Click(object sender, RoutedEventArgs e)
         {
-            using var db = new Wallet();
-            string selected = incomeCategoryList.SelectedItem.ToString();
-            IQueryable<IncomeCategory> categoryToRemove = db.IncomeCategories.Where(el => el.Category == selected);
-            db.IncomeCategories.RemoveRange(categoryToRemove);
-            int result = db.SaveChanges();
-            updateIncomeCategory();
+            if (incomeCategoryList.SelectedItem is not null)
+            {
+                using var db = new Wallet();
+                string selected = incomeCategoryList.SelectedItem.ToString();
+                IQueryable<IncomeCategory> categoryToRemove = db.IncomeCategories.Where(el => el.Category == selected);
+                db.IncomeCategories.RemoveRange(categoryToRemove);
+                int result = db.SaveChanges();
+                updateIncomeCategory();
+            }
         }
         private void removeOutgoingCategory_Click(object sender, RoutedEventArgs e)
         {
-            using var db = new Wallet();
-            string selected = outgoingCategoryList.SelectedItem.ToString();
-            IQueryable<OutgoingsCategory> categoryToRemove = db.OutgoingsCategories.Where(el => el.Category == selected);
-            db.OutgoingsCategories.RemoveRange(categoryToRemove);
-            int result = db.SaveChanges();
-            updateOutgoingCategory();
+            if(outgoingCategoryList.SelectedItem is not null)
+            {
+                using var db = new Wallet();
+                string selected = outgoingCategoryList.SelectedItem.ToString();
+                IQueryable<OutgoingsCategory> categoryToRemove = db.OutgoingsCategories.Where(el => el.Category == selected);
+                db.OutgoingsCategories.RemoveRange(categoryToRemove);
+                int result = db.SaveChanges();
+                updateOutgoingCategory();
+            }
         }
         private void addIncomeCat(string cat)
         {
-            using var db = new Wallet();
-            var newCategory = new IncomeCategory();
-            newCategory.Category = cat;
-            db.IncomeCategories.Add(newCategory);
-            int succes = db.SaveChanges();
-            updateIncomeCategory();
+            if (!string.IsNullOrEmpty(cat))
+            {
+                using var db = new Wallet();
+                var newCategory = new IncomeCategory();
+                newCategory.Category = cat;
+                db.IncomeCategories.Add(newCategory);
+                int succes = db.SaveChanges();
+                updateIncomeCategory();
+            }
         }
         private void addOutgoingCat(string cat)
         {
-            using var db = new Wallet();
-            var newCategory = new OutgoingsCategory();
-            newCategory.Category = cat;
-            db.OutgoingsCategories.Add(newCategory);
-            int succes = db.SaveChanges();
-            updateOutgoingCategory();
+            if(!string.IsNullOrEmpty(cat))
+            {
+                using var db = new Wallet();
+                var newCategory = new OutgoingsCategory();
+                newCategory.Category = cat;
+                db.OutgoingsCategories.Add(newCategory);
+                int succes = db.SaveChanges();
+                updateOutgoingCategory();
+            }
+
         }
         private void addCategory_Click(object sender, RoutedEventArgs e)
         {
